Create sample challenge data in json
Only a subset of the full challenges.xml has been converted - just enough to get a couple of steps into the progression. sortOrder is a new field that for legacy Challenges will be essentially the increasing order of the legacy IDs. The CMS should allow users to re-order challenges, which will be reflected in the sortOrder attribute